requirements.txt
matplotlib>=3.8.4
mitsuba>=3.5.0
numpy>=1.26.4
opencv_python>=4.9.0.80
pandas>=2.2.2
Pillow>=10.3.0
PyYAML>=6.0.1
scipy>=1.13.0
tqdm>=4.66.1
yacs>=0.1.8
